feat: BigDecimal로 값 수정

+++ src/main/java/com/poortorich/chart/constants/ChartConstants.java
@@ -1,11 +1,13 @@
 package com.poortorich.chart.constants;
 
+import java.math.BigDecimal;
+
 public class ChartConstants {
 
     public static final String DUMMY_CATEGORY = "dummy";
     public static final String DUMMY_COLOR = "#e6e6e6";
-    public static final Double DUMMY_RATE = 100.;
-    public static final Double AGGREGATED_DATE_MIN_RATE = 1.5;
+    public static final BigDecimal DUMMY_RATE = BigDecimal.valueOf(100);
+    public static final BigDecimal AGGREGATED_DATE_MIN_RATE = BigDecimal.valueOf(1.5);
 
     private ChartConstants() {
     }
