+++ src/main/java/com/poortorich/chart/response/CategoryChart.java
@@ -1,5 +1,6 @@
 package com.poortorich.chart.response;
 
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Data;
 
@@ -10,6 +11,6 @@ public class CategoryChart {
     private Long id;
     private String color;
     private String name;
-    private Double rate;
+    private BigDecimal rate;
     private Long amount;
 }
